SystemUI: there is unexpected SIM PIN input dialog.
There is unexpected input dialog for SIM PIN.

Because the SystemUI set it sim state to ready after
verified the PIN lock, but the sim state in the
telephony is in PIN_REQUIRED.So the SystemUI will get
another security screen.

Do no act to PIN_REQUIRED after in READY state.

Change-Id: I9bbe5034b48c62c451b0ef6dad0da1da428f6532
CRs-Fixed: 2669805
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Anushek Prasal <anushekprasal@gmail.com>